Script allows to unfollow from all followers in Twitter
It seems to be common thing when a twitter account is hacked and a lot
of useless followers have appeared. There is no magic button "To
Unfollow All". And what should we do if there are really a lot of
followers, for instance, over 1000 ? So, this Javascript code can help.